# Produce better code from maskW