@treeowl 2019-01-22
Good point! That *should* be better. But since you're micro-optimizing here, I think it would pay to get some benchmark results too, to make sure that you don't happen by chance to damage alignment and actually reduce performance. As long as benchmarks either improve or stay the same, I agree we should make this change.