@Fuuzetsu 2019-01-24
I couldn't conclusively confirm which version is better just by benchmarking these separately. I think I'll have to run the containers benchmark and conclude from that.