@treeowl 2019-01-30
Could you zoom in on those `lookup` benchmarks? Stick some other test in front to warm them up a bit, and run them harder. If that's a real regression (even if fairly minor), then I'm a bit concerned.