@treeowl 2019-01-30
For future reference: I personally prefer to see benchmarks woven together: old-new, old-new, old-new, well labeled. Trying to compare by scrolling up and down is tiresome.