@Fuuzetsu 2019-01-31
I think it's noise on my machine.

I added a "warmup" in front (it's just the same benchmark under different name) and I'm running with `-L 60`

without my change
```
benchmarking lookup_warmup
time                 186.0 μs   (185.5 μs .. 186.5 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 185.8 μs   (185.3 μs .. 186.4 μs)
std dev              2.706 μs   (2.027 μs .. 3.571 μs)
variance introduced by outliers: 11% (moderately inflated)

benchmarking lookup
time                 185.9 μs   (185.3 μs .. 186.4 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 185.4 μs   (185.1 μs .. 185.9 μs)
std dev              2.034 μs   (1.506 μs .. 3.355 μs)
```
with my change:
```
benchmarking lookup_warmup
time                 184.2 μs   (183.5 μs .. 184.9 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 183.4 μs   (183.0 μs .. 184.0 μs)
std dev              2.238 μs   (1.905 μs .. 2.833 μs)

benchmarking lookup
time                 185.2 μs   (184.7 μs .. 185.7 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 185.2 μs   (184.6 μs .. 185.9 μs)
std dev              2.974 μs   (2.459 μs .. 3.662 μs)
variance introduced by outliers: 13% (moderately inflated)
```

>For future reference: I personally prefer to see benchmarks woven together: old-new, old-new, old-new, well labeled. Trying to compare by scrolling up and down is tiresome.

I just opened the two side by side, no jumping around needed.